requirements.txt
streamlit==1.32.0
opencv-python==4.8.1.78
mediapipe==0.10.7
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
joblib==1.3.2
pillow==10.0.0
